+++ tests/assets/controllers/rekalogika/rekapager-bundle/infinite-scrolling-controller.js
@@ -1,0 +1,1 @@
+../../../../../packages/rekapager-bundle/assets/dist/infinite-scrolling.js