+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * feat: AssetMapper compatibility
 * fix: js module
 * fix: change infinite scrolling breakpoint to 768px
+* feat: infinite scrolling demo
+* feat: `rekapager_infinite_scrolling_content` Twig function
 
 ## 0.6.2